apu rewrite; buggy;
